feat: connect4 to recoil
